3DVision/requirements.txt
streamlit
transformers
torch
pillow
numpy
opencv-python-headless